backend/requirements.txt
fastapi==0.104.1
jinja2==3.1.3
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.0
python-multipart==0.0.6
asyncpg==0.29.0
apscheduler==3.10.4
requests==2.31.0